Extract ProjectInvoiceLedger out of ProjectFinancePanel
Pure extraction, no behavior change to the Finance tab: the invoice
half of ProjectFinancePanel (menu, list, issue/void/payment/document
dialogs) becomes its own self-contained component, self-querying like
QuoteCard/InvoicingCard/ProjectQuoteRail already do. ProjectFinancePanel
is now a thin composer of ProjectQuoteRail + ProjectInvoiceLedger.

Needed so the invoices manager modal (next commit) can embed the exact
same component the Finance tab uses instead of a second copy.
